+++ clang/lib/Parse/ParseDeclCXX.cpp
@@ -3096,10 +3096,23 @@ Parser::DeclGroupPtrTy Parser::ParseCXXClassMemberDeclaration(
   }
 
   ParsedAttributes DeclSpecAttrs(AttrFactory);
-  MaybeParseMicrosoftAttributes(DeclSpecAttrs);
-
   // Hold late-parsed attributes so we can attach a Decl to them later.
   LateParsedAttrList CommonLateParsedAttrs;
+
+  while (MaybeParseCXX11Attributes(DeclAttrs) ||
+         MaybeParseGNUAttributes(DeclSpecAttrs, &CommonLateParsedAttrs) ||
+         MaybeParseMicrosoftAttributes(DeclSpecAttrs))
+    ;
+
+  SourceLocation DeclStart;
+  if (DeclAttrs.Range.isValid()) {
+    DeclStart = DeclSpecAttrs.Range.isInvalid()
+                    ? DeclAttrs.Range.getBegin()
+                    : std::min(DeclAttrs.Range.getBegin(),
+                               DeclSpecAttrs.Range.getBegin());
+  } else {
+    DeclStart = DeclSpecAttrs.Range.getBegin();
+  }
 
   // decl-specifier-seq:
   // Parse the common declaration-specifiers piece.
@@ -3127,6 +3140,9 @@ Parser::DeclGroupPtrTy Parser::ParseCXXClassMemberDeclaration(
 
   // Turn off colon protection that was set for declspec.
   X.restore();
+
+  if (DeclStart.isValid())
+    DS.SetRangeStart(DeclStart);
 
   // If we had a free-standing type definition with a missing semicolon, we
   // may get this far before the problem becomes obvious.